refactor(api): move controller method
getUserAuthenticationMethods
